requirements.txt
nonebot2
python-dotenv
nonebot-plugin-apscheduler
nonebot-adapter-onebot
psutil
PyYAML
Pillow
numpy
aiohttp
